refactor(cmd): add tar flag
Used for being the source file where the "upload" process adds the files to and then it uploads to the bucket. The "download" process uses it also to download the data to the local filesystem with that filename.

This tar file is used to save every path given during the steps of the task keeping all together in the same tar file. At the end of the task is uploaded then to the bucket.